# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Open Sky Holidays API",
        "description": "API documentation for Open Sky Holidays",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/contact-section/active": {
            "get": {
                "tags": [
                    "Contact Section"
                ],
                "summary": "Get active contact section data",
                "description": "Retrieves the active contact information section for the website contact page.",
                "operationId": "ebffd429abe241c623db5e91a5af6970",
                "responses": {
                    "200": {
                        "description": "Active contact section retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Active contact section retrieved successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ContactSection"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No active contact section found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No active contact section found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/contact-sections/{contactSection}": {
            "get": {
                "tags": [
                    "Contact Section"
                ],
                "summary": "Get a specific contact section",
                "description": "Retrieves a single contact section by its ID.",
                "operationId": "1b6b4e8ffe71cc2c27b9aa4dc0177470",
                "parameters": [
                    {
                        "name": "contactSection",
                        "in": "path",
                        "description": "ID of the contact section",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contact section retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Contact section retrieved successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ContactSection"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Contact section not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Contact section not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/counters/{counter}": {
            "get": {
                "tags": [
                    "Counters"
                ],
                "summary": "Get single statistic counter details",
                "description": "Retrieves single counter details by ID.",
                "operationId": "f393bc04936285b55ba2b81767117aae",
                "parameters": [
                    {
                        "name": "counter",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Counter retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Counter retrieved successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Counter"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Counter not found"
                    }
                }
            }
        },
...
        "/api/heroes/{hero}": {
            "get": {
                "tags": [
                    "Hero Section"
                ],
                "summary": "Get single hero slide details",
                "description": "Retrieves single hero slide details by ID.",
                "operationId": "76559dbbb2604693010df5ee74578b99",
                "parameters": [
                    {
                        "name": "hero",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hero slide retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Hero slide retrieved successfully."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Hero"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Hero slide not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Counter": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "value": {
                        "type": "string",
                        "example": "25+"
                    },
                    "name": {
                        "type": "string",
                        "example": "Years Of Experience"
                    },
                    "icon": {
                        "type": "string",
                        "example": "fa-solid fa-trophy"
                    },
                    "status": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-08-31T10:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-08-31T10:00:00.000000Z"
                    }
                },
                "type": "object"
            },
...
            "Hero": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Unforgettable Journeys Await"
                    },
                    "subtitle": {
                        "type": "string",
                        "example": "Book your dream vacation today"
                    },
                    "image": {
                        "description": "Hero slide image (JPG, JPEG, PNG, WEBP, AVIF)",
                        "type": "string",
                        "example": "hero/slide1.jpg"
                    },
                    "btn_text": {
                        "type": "string",
                        "example": "Explore Packages"
                    },
                    "btn_link": {
                        "type": "string",
                        "example": "/tours"
                    },
                    "status": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
            "ContactSection": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "phone": {
                        "type": "string",
                        "example": "+91 99081 17712"
                    },
                    "email": {
                        "type": "string",
                        "example": "info@openskyholidays.com"
                    },
                    "address": {
                        "type": "string",
                        "example": "#1-11-110, Shyamlal Building, Begumpet, Hyderabad - 500018"
                    },
                    "map_link": {
                        "type": "string",
                        "example": "https://www.google.com/maps/search/?api=1&query=Shyamlal+Building+Begumpet+Hyderabad+500018"
                    },
                    "whatsapp_number": {
                        "type": "string",
                        "example": "+91 99081 17712"
                    },
                    "map_embed_url": {
                        "type": "string",
                        "example": "https://maps.google.com/maps?q=Begumpet&output=embed",
                        "nullable": true
                    },
                    "status": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-08-20T11:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-08-20T11:00:00.000000Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}